Fix compilation error in BondTest
This fixes a compilation error in the BondTest class in
which qCompare<unsigned int, unsigned long> was not found.

+++ tests/auto/chemkit/bond/bondtest.cpp
@@ -47,7 +47,7 @@ void BondTest::atoms()
     chemkit::Bond *bond = molecule.addBond(H1, H2);
     QVERIFY(bond->atom1() == H1);
     QVERIFY(bond->atom2() == H2);
-    QCOMPARE(bond->atoms().size(), 2UL);
+    QCOMPARE(bond->atoms().size(), size_t(2));
     QVERIFY(bond->atoms()[0] == H1);
     QVERIFY(bond->atoms()[1] == H2);
 }
